+++ praw/models/reddit/subreddit.py
@@ -2720,6 +2720,42 @@ class ModeratorRelationship(SubredditRelationship):
         url = API_PATH["friend"].format(subreddit=self.subreddit)
         self.subreddit._reddit.post(url, data=data)
 
+    def invited(self, redditor=None, fetch_all=False, **generator_kwargs):
+        """Return a :class:`.ListingGenerator` for Redditors invited to be moderators.
+
+        :param redditor: When provided, return a list containing at most one
+            :class:`~.Redditor` instance. This is useful to confirm if a relationship
+            exists, or to fetch the metadata associated with a particular relationship
+            (default: None).
+        :param fetch_all: If True, all invited moderators are fetched.
+            (default: False)
+
+            .. note::
+
+                If True, requests will be made until all invited moderators are fetched.
+
+        Additional keyword arguments are passed in the initialization of
+        :class:`.ListingGenerator`.
+
+        .. note::
+
+            Unlike other usages of :class:`.ListingGenerator`, ``limit`` has no effect in
+            the quantity returned. This endpoint always returns moderators in batches of
+            25 at a time regardless of what ``limit`` is set to.
+
+        Usage:
+
+        .. code-block:: python
+
+            for invited_mod in reddit.subreddit("NAME").moderator.invited():
+                print(invited_mod)
+
+        """
+        generator_kwargs["params"] = {"username": redditor} if redditor else None
+        url = API_PATH["list_invited_moderator"].format(subreddit=self.subreddit)
+        generator = ListingGenerator(self.subreddit._reddit, url, **generator_kwargs)
+        return [mod for mod in generator] if fetch_all else generator
+
     def leave(self):
         """Abdicate the moderator position (use with care).